Put hook handler back into sveltekitStrategy
