+++ src/integration.ts
@@ -1,8 +1,10 @@
-import {getTokenStorageType} from "./oauth"
-import type {TokenStorage} from "./oauth/tokenStorage"
-import {browserCookie} from "./oauth/tokenStorage/browserCookie"
-import {localStorage} from "./oauth/tokenStorage/localStorage"
-import { serverCookie } from "./oauth/tokenStorage/serverCookie"
+import { debug } from "svelte/internal"
+
+import { getTokenStorageType } from "./oauth"
+import type { TokenStorage } from "./oauth/tokenStorage"
+import { browserCookie } from "./oauth/tokenStorage/browserCookie"
+import { localStorage } from "./oauth/tokenStorage/localStorage"
+import { getResponseCookie, serverCookie, setRequestCookies } from "./oauth/tokenStorage/serverCookie"
 
 import { browser } from "$app/environment"
 
@@ -25,7 +27,7 @@ export interface ContextStrategy {
      * @param {string} uri The URI of the data
      * @param {Record<string,any>} [options] Fetch options
      */
-    fetch(uri:string, options?:Record<string, unknown>): Promise<Response>,
+    fetch(uri: string, options?: Record<string, unknown>): Promise<Response>,
 
     /**
      * Get the storage where token is saved
@@ -36,7 +38,7 @@ export interface ContextStrategy {
      * Get data from the temporary storage
      * @param {string} key The name/key of the data
      */
-    getFromTemporary(key: string): Promise<string|null>,
+    getFromTemporary(key: string): Promise<string | null>,
 
     /**
      * Save data in the temporary storage
@@ -49,7 +51,7 @@ export interface ContextStrategy {
 export const svelteKitStrategy: ContextStrategy = new class implements ContextStrategy {
     private fetchFunc
     private redirectedTo = null
-    private queryObject: URLSearchParams|null = null
+    private queryObject: URLSearchParams | null = null
 
     fetch(uri: string, options?: Record<string, unknown>): Promise<Response> {
         return this.fetchFunc(uri, options)
@@ -75,12 +77,12 @@ export const svelteKitStrategy: ContextStrategy = new class implements ContextSt
         // const stores = await import("$app/stores")
         // return get(stores.page).query
 
-        // New version, except Page doesn't HAVE a query
+        // New version, except the page store is a Readable and can only be subscribed
         // const page = getStores().page
-        // return page.query
+        // return page.url.searchParams
     }
 
-    getRedirection(): string|null {
+    getRedirection(): string | null {
         const redirection = this.redirectedTo + ""
         this.redirectedTo = null
         return redirection
@@ -118,11 +120,55 @@ export const svelteKitStrategy: ContextStrategy = new class implements ContextSt
 
     async saveInTemporary(key: string, data: string) {
         if (!browser) {
+            debug("integration.ts", "", "", "Saving in in-memory storage")
             inMemoryStorage[key] = data
             return
         }
+        debug("integration.ts", "", "", "Saving in browser session storage")
         return window.sessionStorage.setItem(key, data)
     }
+
+    /**
+     * Handle hooks for SSR
+     * @param {import("@sveltejs/kit/types/hooks").ServerRequest} request The server request
+     * @param {Function} resolve The request resolver
+     */
+    async handleHook({event, resolve}) {
+        debug("integration.ts", "", "", "Handle hook")
+        if (getTokenStorageType() === "cookie" && !browser) {
+            setRequestCookies(event.request.headers["cookie"] || "")
+        }
+
+        const response = await resolve(event)
+
+        return Promise.resolve(response).then((response: Response) => {
+            const cookies = getResponseCookie()
+            debug("integration.ts", "", "", cookies)
+            if (cookies !== "") {
+                let existing = response.headers["set-cookie"] || []
+                if (typeof existing === "string") existing = [existing]
+                existing.push(cookies)
+                response.headers.set("set-cookie", existing)
+            }
+
+            // eslint-disable @typescript-eslint/ban-ts-comment
+            // @ts-ignore: Object is possibly 'undefined'.
+            const redirection = this.getRedirection()
+
+            if (redirection !== null && redirection !== "null") {
+                response = {
+                    ...response,
+                    status: 302,
+                    body: null
+                }
+
+                response.headers.set("location", redirection)
+            }
+
+            return response
+        })
+    }
+
 
 }
 
@@ -131,12 +177,15 @@ export const browserStrategy: ContextStrategy = new class implements ContextStra
         window.location.href = url
         return Promise.resolve()
     }
+
     query(): Promise<URLSearchParams> {
         return Promise.resolve(new URL(window.location.href).searchParams)
     }
+
     fetch(uri: string, options?: Record<string, unknown>): Promise<Response> {
         return fetch(uri, options)
     }
+
     tokenStorage(): Promise<TokenStorage> {
         if (getTokenStorageType() === "cookie") {
             return Promise.resolve(browserCookie)
