+++ src/oauth/grant.ts
@@ -32,15 +32,17 @@ export abstract class BaseGrant implements Grant {
     }
 
     protected getToken(params: Record<string, unknown>, headers: HeadersInit = {}): Promise<boolean> {
-        const requestHeader = new Headers(headers)
-        requestHeader.set("content-type", "application/json")
+        const requestHeaders = new Headers()
+        requestHeaders.set("content-type", "application/json")
 
         return this.integration.fetch(this.tokenUri, {
             method: "post",
             body: JSON.stringify(params),
-            headers: requestHeader
+            headers: requestHeaders
         })
-            .then((response) => response.json())
+            .then((response) => {
+                return response.json()
+            })
             .then(async (response) => {
                 if (Object.keys(response).includes("error")) {
                     (await this.integration.tokenStorage()).set(null)
